requirements.txt
rembg>=2.0.0
pillow>=10.0.0
onnxruntime>=1.15.0
customtkinter>=5.2.0
